split agents into multiple files
